requirements.txt
audiocraft
numpy
simpleaudio
pytaglib
seewav